Make API endpoint configurable
Instead of setting a constant in the class, this is now accepted in the
ZeroPush::Client constructor. The URL from the constant is used as the
default parameter to maintain backwards compatibility.

I'm not a fan of having attr_accessors just to use the setter in the
initialize method, but I simply followed the style in the code.

There is still no way to configure the URL and keep using the
ZeroPush.notify delegate methods so you would have to create a Client
instance yourself - I don't think this is an issue for now.